XMLHttpRequest/send-authentication-competing-names-passwords.htm

<!doctype html>
<html>
  <head>
    <title>XMLHttpRequest: send() - "Basic" authenticated requests with competing user name/password options</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-open()-method" data-tested-assertations="following::ol[1]/li[9]/ol[1]/li[1] following::ol[1]/li[9]/ol[1]/li[2]" />
    <link rel="help" href="http://dvcs.w3.org/hg/xhr/raw-file/tip/Overview.html#the-send()-method" data-tested-assertations="following::code[contains(@title,'http-authorization')]/.." />  </head>
  <body>
    <div id="log"></div>
    <script>
      function request(user1, pass1, user2, pass2, shouldWaitForChallenge, name) { 
        // user1, pass1 will if given become userinfo part of URL
        // user2, pass2 will if given be passed to open() call
        test(function() {
          var client = new XMLHttpRequest(),
              urlstart = "", userwin, passwin
          // if user2 is set, winning user name and password is 2
          if(user2)
            userwin = user2, passwin = pass2
          // if user1 is set, and user2 is not set, user1 and pass1 win
          if(user1 && ! user2)
            userwin = user1, passwin = pass1
          // if neither user name is set, pass 2 wins (there will be no userinfo in URL)
          if (!(user1 || user2))
            passwin = pass2
          if(user1) { // should add userinfo to URL (there is no way to create userinfo part of URL with only password in)
            urlstart = "http://" + user1
            if(pass1)
              urlstart += ":" + pass1
            urlstart += "@" + location.host + location.pathname.replace(/\/[^\/]*$/, '/')
          }
          client.open("GET", urlstart + "resources/authentication.php", false, user2, pass2)
          client.setRequestHeader("x-user", userwin)
          client.setRequestHeader("x-pass", passwin)
          client.send(null)
          assert_equals(client.responseText, (userwin||'') + "\n" + (passwin||''))
          assert_equals(client.getResponseHeader('X-challenge'), shouldWaitForChallenge ? 'DID' : 'DID-NOT' )
          // We want to send multiple requests to the same realm here, so we try to make the UA forget its (cached) credentials between each test..
          // forcing a 401 response to (hopefully) "log out"
          // NOTE: This is commented out because it causes authentication prompts while running the test
          //client.open('GET', "resources/authentication.php?logout=1", false)
          //client.send()
        }, document.title+' '+name)
      }
      var ident = Math.random()
      request(null, null, ident+1, ident+2, true, 'user/pass in open() call')
      request(null, null, ident+3, ident+4, false, 'another user/pass in open() call - must override cached credentials from previous test')
      request("userinfo-user", "userinfo-pass", ident+5, ident+6, false, 'user/pass both in URL userinfo AND open() call - expexted that open() wins')
      request(ident+7, ident+8, null, null, false, 'user/pass *only* in URL userinfo')
      request(ident+9, null, null, ident+10, false, 'user name in URL userinfo, password in open() call: user name wins and password is thrown away')
      request("1", ident+11, ident+12, null, false, 'user name and password in URL userinfo, only user name in open() call: user name in open() wins')
    </script>
  </body>
</html>
